requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.4
django-rest-framework==0.1.0
djangorestframework==3.8.2
idna==2.6
pytz==2018.4
requests==2.18.4
urllib3==1.22
v-vk-api==1.3
